+++ package.json
@@ -35,6 +35,8 @@
     "jsonwebtoken": "^8.5.1",
     "multer": "^1.4.2",
     "pg": "^8.2.1",
+    "reflect-metadata": "^0.1.13",
+    "tsyringe": "^4.3.0",
     "typeorm": "^0.2.25",
     "uuidv4": "^6.0.8"
   }
