+++ src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,23 +1,10 @@
 import { Router } from 'express';
 
-import CreateSessionService from '@modules/users/services/CreateSessionService';
-import UsersRepository from '../../typeorm/repositories/UsersRepository';
+import SessionsController from '../controllers/SessionsController';
 
 const sessionRouter = Router();
+const sessionsController = new SessionsController();
 
-sessionRouter.post('/', async (request, response) => {
-  const usersRepository = new UsersRepository();
-  const { email, password } = request.body;
-  const createSession = new CreateSessionService(usersRepository);
-
-  const { user, token } = await createSession.execute({
-    email,
-    password,
-  });
-
-  delete user.password;
-
-  return response.json({ user, token });
-});
+sessionRouter.post('/', sessionsController.create);
 
 export default sessionRouter;
