+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -1,55 +1,27 @@
 import { Router } from 'express';
 import multer from 'multer';
 
-import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
-import CreateUserService from '@modules/users/services/CreateUserService';
+import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
 
 import uploadConfig from '@config/upload';
+import UsersController from '../controllers/UsersController';
+import UserAvatarController from '../controllers/UserAvatarController';
 
-import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
-import UsersRepository from '../../typeorm/repositories/UsersRepository';
+// ---------------------------------------------------------------------------//
 
 const usersRouter = Router();
+const usersController = new UsersController();
+const userAvatarController = new UserAvatarController();
 
 const upload = multer(uploadConfig);
 
-usersRouter.post('/', async (request, response) => {
-  const usersRepository = new UsersRepository();
-  try {
-    const { email, name, password } = request.body;
-
-    const createUser = new CreateUserService(usersRepository);
-
-    const user = await createUser.execute({ name, email, password });
-
-    delete user.password;
-
-    return response.json(user);
-  } catch (err) {
-    return response.status(400).json({ error: err.message });
-  }
-});
+usersRouter.post('/', usersController.create);
 
 usersRouter.patch(
   '/avatar',
   ensureAuthenticated,
   upload.single('avatar'),
-  async (request, response) => {
-    const usersRepository = new UsersRepository();
-
-    const updateUserAvatarService = new UpdateUserAvatarService(
-      usersRepository,
-    );
-
-    const user = await updateUserAvatarService.execute({
-      user_id: request.user.id,
-      avatarFilename: request.file.filename,
-    });
-
-    delete user.password;
-
-    return response.json(user);
-  },
+  userAvatarController.update,
 );
 
 export default usersRouter;
